requirements.txt
djangorestframework
django-filter
django-crispy-forms
django_enumfield
dateutils
